feat(builder): :sparkles: Introduce transaction based tree updates
General Update
Instead of sending entire trees at once a patch is generated on each client tree update. This patch is then synced with the backend, which applies it to its local tree.

Sync
The updated syncMachine handles syncing and, in case of errors, the error. Patches are saved locally in addition to the tree and are reloaded when the app is reinitialized.

Undo/Redo
Based on the patches undo redo has been added.

Refactors
- the context of the treeMachine does no have a key of tree with the tree; the tree is not spread in the context anymore
  - this change was made, because more properties have been added to the context that are not part of the tree directly