feat: Finalize Rework
Combined with the previous 9 commits this one wraps 3 months of work on Open Decision.
We reimagined what it means to build a decision tree to be far grander then
just decisions. Document generation, Forms, Variables, Modules, Theming
all make Open Decision more of a tool builder.

While not all of these features are stable we want to more incrementally
improve the project instead of continuing on this monster of a branch.