feat(builder): Introduce separate sharing page for prototype
In order to better distinguish between the authenticated and
unauthenticated state of the prototype a separate shared page has been
introduced. This page is only accessible when the prototype is shared
by setting the hasPreview flag. If the prototype is visited by an
unauthenticated user, the user is redirected to the login page.

Includes:
- refactor of api client
- refactor of react-binding for api client
- addition of shared prototype page